refactor stats service
