+++ submission/submission-core/src/main/kotlin/ac/uk/ebi/biostd/submission/stats/SubmissionStatsService.kt
@@ -7,19 +7,17 @@ import ac.uk.ebi.biostd.persistence.common.request.PageRequest
 import ac.uk.ebi.biostd.persistence.common.service.StatsDataService
 import ac.uk.ebi.biostd.persistence.common.service.SubmissionPersistenceQueryService
 import ac.uk.ebi.biostd.persistence.doc.model.SingleSubmissionStat
-import ac.uk.ebi.biostd.submission.helpers.TempFileGenerator
 import ebi.ac.uk.extended.model.ExtSubmission
 import kotlinx.coroutines.flow.Flow
 import mu.KotlinLogging
-import org.springframework.web.multipart.MultipartFile
 import uk.ac.ebi.extended.serialization.service.ExtSerializationService
 import uk.ac.ebi.extended.serialization.service.filesFlow
+import java.io.File
 
 private val logger = KotlinLogging.logger {}
 
 class SubmissionStatsService(
     private val statsFileHandler: StatsFileHandler,
-    private val tempFileGenerator: TempFileGenerator,
     private val submissionStatsService: StatsDataService,
     private val serializationService: ExtSerializationService,
     private val extSubmissionQueryService: SubmissionPersistenceQueryService,
@@ -43,17 +41,13 @@ class SubmissionStatsService(
         stat: SubmissionStat,
     ): SubmissionStat = submissionStatsService.save(stat)
 
-    suspend fun register(type: String, stats: MultipartFile): List<SubmissionStat> {
-        val statsFile = tempFileGenerator.asFile(stats)
-        val statsList = statsFileHandler.readStats(statsFile, SubmissionStatType.fromString(type.uppercase()))
-
+    suspend fun register(type: String, stats: File): List<SubmissionStat> {
+        val statsList = statsFileHandler.readStats(stats, SubmissionStatType.fromString(type.uppercase()))
         return submissionStatsService.saveAll(statsList)
     }
 
-    suspend fun increment(type: String, stats: MultipartFile): List<SubmissionStat> {
-        val statsFile = tempFileGenerator.asFile(stats)
+    suspend fun increment(type: String, statsFile: File): List<SubmissionStat> {
         val statsList = statsFileHandler.readStats(statsFile, SubmissionStatType.fromString(type.uppercase()))
-
         return submissionStatsService.incrementAll(statsList)
     }
