+++ submission/submission-webapp/src/main/kotlin/ac/uk/ebi/biostd/common/config/StatsConfig.kt
@@ -1,8 +1,9 @@
 package ac.uk.ebi.biostd.common.config
 
+import ac.uk.ebi.biostd.common.properties.ApplicationProperties
 import ac.uk.ebi.biostd.persistence.common.service.StatsDataService
 import ac.uk.ebi.biostd.persistence.common.service.SubmissionPersistenceQueryService
-import ac.uk.ebi.biostd.submission.helpers.TempFileGenerator
+import ac.uk.ebi.biostd.stats.web.TempFileGenerator
 import ac.uk.ebi.biostd.submission.stats.StatsFileHandler
 import ac.uk.ebi.biostd.submission.stats.SubmissionStatsService
 import org.springframework.context.annotation.Bean
@@ -15,15 +16,17 @@ class StatsConfig {
     fun statsFileHandler(): StatsFileHandler = StatsFileHandler()
 
     @Bean
+    fun tempFileGenerator(properties: ApplicationProperties): TempFileGenerator =
+        TempFileGenerator(properties.tempDirPath)
+
+    @Bean
     fun submissionStatsService(
         statsFileHandler: StatsFileHandler,
-        tempFileGenerator: TempFileGenerator,
         submissionStatsService: StatsDataService,
         extSerializationService: ExtSerializationService,
         pesistenceQueryService: SubmissionPersistenceQueryService,
     ): SubmissionStatsService = SubmissionStatsService(
         statsFileHandler,
-        tempFileGenerator,
         submissionStatsService,
         extSerializationService,
         pesistenceQueryService,
