+++ submission/submission-webapp/src/test/kotlin/ac/uk/ebi/biostd/submission/domain/service/TempFileGeneratorTest.kt
@@ -1,6 +1,6 @@
 package ac.uk.ebi.biostd.submission.domain.service
 
-import ac.uk.ebi.biostd.submission.helpers.TempFileGenerator
+import ac.uk.ebi.biostd.stats.web.TempFileGenerator
 import io.github.glytching.junit.extension.folder.TemporaryFolder
 import io.github.glytching.junit.extension.folder.TemporaryFolderExtension
 import io.mockk.junit5.MockKExtension
